refactor(chat): extract WaveformTimeline into its own file to fit the 500-line gate
frontend/features/chat/components/ChatComposerControls.tsx was 503
lines, failing scripts/check-file-lines.mjs and the bun run check
gate.  WaveformTimeline (the recording-state scrolling bar widget) and
its WAVEFORM_BARS constant are a self-contained visual subcomponent
with no shared state, so they extract cleanly to their own file —
ChatComposerControls.tsx drops to 444 lines.

VoiceMeter still owns the WaveformTimeline render; the only edit on
the consumer side is the new import line.

Includes the bean (.beans/ai-nexus-gzk8) that tracked all three CI
fixes landing on this branch.
